Transfer into multiple-project build and support 1.18.x
(cherry picked from commit afe76def1821feb686756d53f1da801d6863d4e5)
(cherry picked from commit eb235163864e93516f7db9149eefc5543c76cf70)
