+++ 1_18_R2/build.gradle.kts
@@ -1,0 +1,21 @@
+plugins {
+    `java-library`
+    id("io.papermc.paperweight.userdev") version "1.3.6"
+    id("xyz.jpenilla.run-paper") version "1.0.6"
+}
+dependencies {
+    paperDevBundle("1.18.2-R0.1-SNAPSHOT")
+    implementation(project(":common"))
+}
+tasks {
+    compileJava {
+        options.encoding = Charsets.UTF_8.name()
+        options.release.set(17)
+    }
+    javadoc {
+        options.encoding = Charsets.UTF_8.name()
+    }
+    processResources {
+        filteringCharset = Charsets.UTF_8.name()
+    }
+}
