+++ common/src/main/java/lordpipe/cleanupstupidmojangshit/NBTRemover.java
@@ -1,0 +1,11 @@
+package lordpipe.cleanupstupidmojangshit;
+
+import org.bukkit.block.Block;
+
+public interface NBTRemover {
+    void removeFilteredTextTag(Block block);
+
+    default boolean isSupported() {
+        return false;
+    }
+}
