+++ settings.gradle.kts
@@ -5,4 +5,7 @@ pluginManagement {
     }
 }
 
+include("common")
+include("1_18_R1")
+include("1_18_R2")
 rootProject.name = "CleanUpStupidMojangShit"
